Add IE7 support for fontawesome icon font.
Adds the IE7 mixin and the individual icon inclusions in each
icon mixin. Prefixes IE7 support inclusions by the .ie7 class, so
that will have to be defined for this to work.

Also note that IE7 support changes the way the icons are rendered
due to the inability to use the :before pseudo-element. Care
needs to be taken when implementing this to ensure that content
is not overwritten by font icon because its being applied to the
selector's element, not its :before pseudo-element.